[MetaSchedule][Minor] Fix Integer Overflow in Tuning Statistics
This PR fixes the integer overflow when the flop count of a given workload is larger than `MAX_INT` during tuning statistics printing.

+++ src/meta_schedule/task_scheduler/gradient_based.cc
@@ -77,7 +77,7 @@ class GradientBasedNode final : public TaskSchedulerNode {
       int trials = record.trials;
       row << /*id=*/i                                 //
           << /*name=*/record.task->task_name.value()  //
-          << /*flops=*/static_cast<int>(record.flop)  //
+          << /*flops=*/static_cast<int64_t>(record.flop)  //
           << /*weight=*/static_cast<int>(record.weight);
       if (trials == 0) {
         row << /*speed=*/"N/A" << /*latency=*/"N/A" << /*weighted_latency=*/"N/A";
